Added Enabled/Disabled Status Support in the 'poll' Function
This commit adds DriveStatus to the status items queried from the
controller. The Automation1DriveStatus_Enabled mask is applied to the
queried item and the resulting bit is used to set the enabled status of
the axis.

Additionally, it was found that ProgramPositionCommand was no longer
being used, so it was removed from the queried items.

The driver was rebuilt and tested. It seems to be working properly.